cfb: Non-octal display height bug fix
This frame buffer implementation only support 8 pixel per tile (ppt set to 8U) but it allow display height (y_res) to be any value without properly adjusting the frame buffer size in cases where the y_res is not devisable by 8 (non-octal). This patch properly size the frame buffer in these cases. For example if y_res is 9 then it assumes there are two 8-bit tiles in the frame buffer.